modal_backend/requirements.txt
modal>=0.73.0
supabase==2.10.0
httpx==0.27.2
pdfplumber==0.11.4
openpyxl==3.1.5
lightrag-hku==1.3.5
feedparser==6.0.11
numpy==1.26.4
